Add persistence test and clean up old script
